ecore - ecore pipe - don't use ecores main loop select except on win
this select could be something that conflicts with the regular main
loop... using this somewhere else like a thread just to listen to that
ecore pipe actually breaks for select. as ecore_pipe_wait waits only
for that pipe then using the min loop select is probelmatic, so use
regular select except in windows where we have our own wrapper.

@fix